% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/across-sites.R
\name{graph_sensor}
\alias{graph_sensor}
\title{graph sensor data files}
\usage{
graph_sensor(
  x = read_sensor(),
  sensor = c("temp", "PAR", "waves", "current")[1],
  ...
)
}
\arguments{
\item{x}{tibble of single data type}

\item{sensor}{character, type of sensor to be analyzed, "temp", "PAR", "waves", "current"}

\item{...}{further arguments passed to \code{\link[ggplot2]{theme}}}
}
\value{
ggplot
}
\description{
graph sensor data files
}
