% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/across-sites.R
\name{read_sensor}
\alias{read_sensor}
\title{read sensor data files}
\usage{
read_sensor(
  filenames = example_filenames(),
  sensor = c("temp", "PAR", "waves", "current")[1],
  raw = FALSE,
  ...
)
}
\arguments{
\item{filenames}{vector, the name of the files to be read in}

\item{sensor}{character, type of sensor to be analyzed, "temp", "PAR", "waves", "current"}

\item{raw}{logical, FALSE if passing QAQC'd data (default)}

\item{...}{further arguments passed to \code{\link[hobotemp]{theme}}}
}
\value{
tibble
}
\description{
read sensor data files
}
